+++ internal/component/otelcol/connector/host_info/host_info.go
@@ -35,6 +35,9 @@ type Arguments struct {
 
 	// Output configures where to send processed data. Required.
 	Output *otelcol.ConsumerArguments `alloy:"output,block"`
+
+	// DebugMetrics configures component internal metrics. Optional.
+	DebugMetrics otelcol.DebugMetricsArguments `alloy:"debug_metrics,block,optional"`
 }
 
 var (
@@ -49,6 +52,7 @@ func (args *Arguments) SetToDefault() {
 		HostIdentifiers:      []string{"host.id"},
 		MetricsFlushInterval: 60 * time.Second,
 	}
+	args.DebugMetrics.SetToDefault()
 }
 
 // Validate implements syntax.Validator.
@@ -90,4 +94,9 @@ func (args Arguments) NextConsumers() *otelcol.ConsumerArguments {
 // ConnectorType() int implements connector.Arguments.
 func (Arguments) ConnectorType() int {
 	return connector.ConnectorTracesToMetrics
+}
+
+// DebugMetricsConfig implements connector.Arguments.
+func (args Arguments) DebugMetricsConfig() otelcol.DebugMetricsArguments {
+	return args.DebugMetrics
 }